Assume a panel on every monitor
Unity puts a panel on every monitor. The panels on secondary monitors are not
contained in the work area because gdk_screen_get_monitor_workarea() returns
the full geometry for secondary monitors. Synchronous bubbles might thus overlap
the panels on secondary monitors.

Fix that by assuming a panel with the primary panel's height on every monitor.