Refactor bubble_set_icon()
This function did the same as bubble_set_icon_for_path(), because paths
are allowed in addition to icon names for both the app-icon parameter
and the image-path hint.

Remove one of those functions and inline load_icon(), which gives us
more control over when icon-not-found warnings are shown. Don't show one
when using the fallback of prefixing the icon name with "notification-".