Listen to xsettings dpi changes
We only listened to font scale factor changes on gsettings, but
retrieved the actual value from xsettings (through gtk-xft-dpi). This is
racy: unity-settings-daemon also listens to gsettings and might not have
set the xsettings value when our handler is run.

Fix this by listening to xsettings changes as well.

This means that we're now listening for both xsettings and gsettings
changes for what is essentially the same thing and recalculate size on
each of them. This creates some unnecessary work, but those settings are
not changed very often. We can't only listen on xsettings, because there
are other gsettings keys that might trigger a change in notification
bubble appearance.